mocks for endpoint dao moved into separate file (mock_daos.go -> mock_dao_endpoint.go)
